refactor: Answer 도메인으로 패키지 변경
