+++ src/test/java/blackjack/domain/AnswerTest.java
@@ -1,4 +1,4 @@
-package blackjack.controller;
+package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
